# backend/src/main/resources/api.yaml
swagger: '2.0'
info:
  description: >-
    This is the API for internal communication between the MeetNow app on mobile
    devices and the backend
  version: 1.0.0
  title: MeetNow Internal Communication API
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
host: meetnow.cfapps.eu10.hana.ondemand.com
basePath: /
schemes:
  - https
paths:
  /newUser:
    post:
      tags: 
        - setupCalls
      summary: make urself known to the backend
      operationId: newUser
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - in: formData
          name: pushId
          type: string
          required: true
          description: the oneSignal push id
      responses:
        '200':
          description: user created, returns id and secret to store
          schema:
            $ref: '#/definitions/User'
  /meeting:
    post:
      tags:
        - meeting
      summary: Submit a new meeting to the backends
      operationId: addMeeting
      consumes:
        - application/json
      produces:
        - text/plain
      parameters:
        - in: body
          name: meeting
          description: Meeting that needs to be submitted to the backend
          required: true
          schema:
            $ref: '#/definitions/Meeting'
      responses:
        '201':
          description: Creation successfull
          schema:
            type: string
            description: the id of the newly created meeting
        '405':
          description: Validation Exception
#    get:
#      tags:
#        - meeting
#      summary: Check whether a meeting exists
#      operationId: getMeeting
#      consumes:
#        - application/json
#      parameters:
#        - in: body
#          name: id
#          description: ID of Meeting which shall exist
#          required: true
#          type: string
#      responses:
#        '200':
#          description: OK
#        '404':
#          description: Meeting id not found
    put:
      tags:
        - meeting
      summary: Update an existing meeting
      operationId: updateMeeting
      consumes:
        - application/json
      parameters:
        - in: body
          name: meeting
          description: Meeting which shall be updated
          required: true
          schema:
            $ref: '#/definitions/Meeting'
      responses:
        '200':
          description: Update successfull
        '401':
          description: 'Not Authorized, if you''re trying to update a meeting you don''t own'
        '404':
          description: Meeting id not found
        '405':
          description: Validation Exception
    delete:
      tags:
        - meeting
      summary: Remove an existing meeting
      operationId: removeMeeting
      consumes:
        - application/json
      parameters:
        - in: body
          name: meeting
          description: Meeting which shall be deleted
          required: true
          schema:
            $ref: '#/definitions/Meeting'
      responses:
        '200':
          description: Deletion successfull
        '401':
          description: 'Not Authorized, if you''re trying to delete a meeting you don''t own'
        '404':
          description: Meeting id not found
        '405':
          description: Validation Exception
  /enterArea:
    post:
      tags:
        - area
      summary: notify the backend that the client enters a defined area
      operationId: enterArea
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - in: formData
          name: meetingId
          description: the id of the meeting which has an area which was entered
          required: true
          type: string
        - in: formData
          name: userId
          description: the (OneSignal) id of the user
          required: true
          type: string
      responses:
        '200':
          description: Update successful
        '400':
          description: Meeting not recognized
  /leaveArea:
    post:
      tags:
        - area
      summary: notify the backend that the client leaves a defined area
      operationId: leaveArea
      consumes:
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - in: formData
          name: meetingId
          description: the id of the meeting which area was left
          required: true
          type: string
        - in: formData
          name: userId
          description: the (OneSignal) id of the user
          required: true
          type: string
      responses:
        '200':
          description: Update successful
        '405':
          description: Not in an area at the moment
        '400':
          description: Meeting not recognized
definitions:
  Meeting:
    type: object
    required:
      - reoccurrence
      - ownerId
    properties:
      id:
        description: unique identifier, generated by server
        type: string
      name:
        type: string
      ownerId:
        type: string
      category:
        description: the meeting category
        type: string
        enum:
          - jourfix
          - lunch
          - coffeebreak
      duration:
        description: how long will the meeting take? (in minutes)
        type: integer
      reoccurrence:
        description: how often shall the meeting reoccur?
        type: string
        enum:
          - daily
          - weekly
          - monthly
      participants:
        type: array
        items:
          $ref: '#/definitions/Participant'
      groups:
        type: array
        items:
          $ref: '#/definitions/Group'
#      nextScheduled:
#        type: string
#        format: date-time
      area:
        $ref: '#/definitions/Area'
  Group:
    type: object
    properties:
      id:
        description: unique identifier
        type: string
      participants:
        type: array
        items:
          $ref: '#/definitions/Participant'
  Participant:
    type: object
    properties:
      name:
        description: full name of the participant
        type: string
      id:
        description: oneSignalId for push Notifications
        type: string
#      busyTimes:
#        type: array
#        items:
#          properties:
#            startTime:
#              type: string
#              format: time
#            duration:
#              type: string
#              format: time
  Area:
    type: object
    properties:
      longitude:
        type: number
      latitude:
        type: number
      radius:
        type: integer
        format: int32
        description: the radius of the area in meters
  User:
    type: object
    properties:
      id:
        type: string
        description: unique id  (OneSignal)
      secret:
        type: string
        description: random string for authentification generated by server